fix (when initailly there is no internet connection)
when you have an internet connection and you launch the app. it's showing connected with the internet.
but when you turn off all networks and lanch the app again it's showing noting. the network callback did not trigger at all until you again turn on the wifi/internet.

fix this issue.